solve editing in group.js
